// app/api/interface/response/get_user_detail.ts
export interface GetUserDetailResponse {
    user_id: string;
    first_name: string;
    last_name: string;
    address: string;
    email: string;
    cccd: string;
    major_names: string[];
    phone_number: string;
    role: string;
    is_priority: boolean;
    is_first_login: boolean;
    wallet_address: string;
    major_ids: string[];
    created_at: string;
    updated_at: string;
    student_code: string;
}